+++ next-cloudinary/tests/lib/cloudinary.spec.js
@@ -1,4 +1,5 @@
 import { getCldImageUrl } from '../../src/helpers/getCldImageUrl';
+import { getCldVideoUrl } from '../../src/helpers/getCldVideoUrl';
 
 describe('Cloudinary', () => {
   const OLD_ENV = process.env;
@@ -12,7 +13,7 @@ describe('Cloudinary', () => {
     process.env = OLD_ENV;
   });
 
-  describe('getCldImage', () => {
+  describe('getCldImageUrl', () => {
     it('should pass', () => {
       const cloudName = 'customtestcloud';
 
@@ -25,6 +26,22 @@ describe('Cloudinary', () => {
       });
 
       expect(url).toContain(`https://res.cloudinary.com/${cloudName}/image/upload/c_limit,w_100/f_auto/q_auto/turtle`);
+    });
+  });
+
+  describe('getCldVideoUrl', () => {
+    it('should pass', () => {
+      const cloudName = 'customtestcloud';
+
+      process.env.NEXT_PUBLIC_CLOUDINARY_CLOUD_NAME = cloudName;
+
+      const url = getCldVideoUrl({
+        src: 'turtle',
+        width: 100,
+        height: 100
+      });
+
+      expect(url).toContain(`https://res.cloudinary.com/${cloudName}/video/upload/c_limit,w_100/f_auto/q_auto/turtle`);
     });
   });
 })